split rust side into multiple files; cleanup headers
