+++ vortex-cxx/cpp/include/vortex/scan.hpp
@@ -7,7 +7,6 @@
 
 #include <cstdint>
 #include <memory>
-#include <utility>
 
 namespace vortex {
