+++ vortex-cxx/cpp/include/vortex/write_options.hpp
@@ -4,9 +4,7 @@
 #pragma once
 
 #include <nanoarrow/common/inline_types.h>
-
 #include <memory>
-#include <string>
 
 namespace vortex {
