+++ vortex-cxx/cpp/src/scan.cpp
@@ -4,6 +4,7 @@
 #include "vortex/scan.hpp"
 #include "vortex/exception.hpp"
 #include "scan_impl.hpp"
+#include "rust/cxx.h"
 
 namespace vortex {
