+++ vortex-cxx/src/write.rs
@@ -1,0 +1,75 @@
+use arrow_array::RecordBatchReader;
+use arrow_array::ffi_stream::{ArrowArrayStreamReader, FFI_ArrowArrayStream};
+use vortex::ArrayRef;
+use vortex::arrow::FromArrowArray;
+use vortex::dtype::DType;
+use vortex::dtype::arrow::FromArrowType;
+use vortex::error::VortexError;
+use vortex::file::VortexWriteOptions as WriteOptions;
+use vortex::iter::{ArrayIteratorAdapter, ArrayIteratorExt};
+use vortex::stream::ArrayStream;
+
+use crate::RUNTIME;
+
+#[cxx::bridge(namespace = "vortex::ffi")]
+mod ffi {
+    extern "Rust" {
+        type VortexWriteOptions;
+        fn write_options_new() -> Box<VortexWriteOptions>;
+        unsafe fn write_array_stream(
+            options: Box<VortexWriteOptions>,
+            input_stream: *mut u8,
+            path: &str,
+        ) -> Result<()>;
+    }
+}
+
+struct VortexWriteOptions {
+    inner: WriteOptions,
+}
+
+fn write_options_new() -> Box<VortexWriteOptions> {
+    Box::new(VortexWriteOptions {
+        inner: WriteOptions::default(),
+    })
+}
+
+/// Convert an ArrowArrayStreamReader to a Vortex ArrayStream
+fn arrow_stream_to_vortex_stream(
+    reader: ArrowArrayStreamReader,
+) -> Result<impl ArrayStream, Box<dyn std::error::Error + Send + Sync>> {
+    let array_iter = ArrayIteratorAdapter::new(
+        DType::from_arrow(reader.schema()),
+        reader.map(|result| {
+            result
+                .map(|record_batch| ArrayRef::from_arrow(record_batch, false))
+                .map_err(VortexError::from)
+        }),
+    );
+
+    Ok(array_iter.into_array_stream())
+}
+
+/// # Safety
+///
+/// input_stream should be valid FFI_ArrowArrayStream.
+/// See [`FFI_ArrowArrayStream::from_raw`]
+unsafe fn write_array_stream(
+    options: Box<VortexWriteOptions>,
+    input_stream: *mut u8,
+    path: &str,
+) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
+    let path = path.to_string();
+
+    let stream_reader =
+        unsafe { ArrowArrayStreamReader::from_raw(input_stream as *mut FFI_ArrowArrayStream) }?;
+
+    let vortex_stream = arrow_stream_to_vortex_stream(stream_reader)?;
+
+    RUNTIME.block_on(async {
+        let file = tokio::fs::File::create(path).await?;
+
+        options.inner.write(file, vortex_stream).await?;
+        Ok::<(), Box<dyn std::error::Error + Send + Sync>>(())
+    })
+}
